ofi: Reduce the size of AM headers
This commit reduces the size of each AM header from 40 bytes to 8 bytes
for shorts and mediums and 16 bytes for longs. This is done through some
bitfields, tranferring the data payload size through immediate data, and
using a union for different types of messages to avoid unneeded fields
being included for shorts and mediums.