Split gasnetc_attach() into multiple pieces
This commit splits gasnetc_attach() into three distinct pieces:

  + gasnet_attach_primary()
  + gasnet_attach_segment()
  + gasnet_attach()

The first contains most of the attach code up to the "primary attach
complete" verbose-debug message, plus a few boiler-plate calls that
follow.  It excludes any code related to registration of client AM
handlers or attaching of the client segment.

The second contains the code the attach/register a client segment.
While currently called exactly once from gasnetc_attach(), it is
intended that this code become the basis for collective segment
creation both within legacy gasnet_attach() and the EX interface
"TeamSegmentCreate".  Note: The *aux* segment is attached earlier
in gasnetc_init().

The remaining gasnet_attach() is a small shell of its former self,
consisting almost entirely of boilerplate such as argument checking,
calls to the other two attach functions, and registration of the
client handlers.  This is a good candidate for future factorization.
However, there are conduit-specific differences to be considered, such
as the need to in udp- and mpi-conduits to hold a lock when making the
call for AM handler registration.