bug 3521: teach testmpi to init thread support as needed
On PAR configurations, testmpi requires either MPI_THREAD_SERIALIZED
or MPI_THREAD_MULTIPLE (depending on the conduit) to run correctly,
so use MPI-2 thread init to enforce this requirement.

Note that due to interoperability issues, testmpi is only guaranteed
to work on mpi-conduit, or systems where the conduit and MPI
implementations are tightly coupled.