aries/gemini: continue to refine split of attach
This commit further refines the gemini/aries conduit's division of
gasnet_attach(), which was begun in the previous commit.

Most significantly, a number of initializations which were previously
contained within gasnetc_attach_segment() were responsible for the
initialization of resources within the aux segment, not the client
segment.  Those have been split-up and are now invoked from within
gasnetc_attach_primary().

This commit rearranges how/when domains other than the first are
initialized (when multi-domain support is enabled).  Since in EX it
will become possible to operate without a segment, much of the domain
initialization has moved to gasnetc_attach_primary(), with only a
small amount remaining at attach time.