Implement gasnetex_ClientInit()
This commit implements gasnetex_ClientInit() in all EX-supported
conduits and the template.  It makes use of the subdivisions of
gasnet_attach() made in previous commits.

While this new function is mostly boilerplate (and thus a good
candidate for conduit-independent factorization), there are still
variations among conduits that will need to be ironed out.  For now
the goal is to get the implementation completed.

The implementation of gasnetex_ClientInit() is now the basis of
gasnet_init().  So, gasnet_init() and gasnet_atatch() are now just
inline "thunks" in gasnet.h.

The name-shifting previously applied to gasnet_init() for link-time
sanity checking is now applied to gasnetex_ClientInit() instead.