Implement gasnetex_TeamSegmentCreate()
This commit utilizes the gasnetc_attach_segment() call (split off from
the main body of gasnetc_attach() in previous commits) to implement
the new gasnetex_TeamSegmentCreate() call in all EX-supported conduits
and the template.

There are limitations to the current implementation, such as:

  + No support for client-supplied segment
  + May call at most once (and gasnet_attach() counts as one call)
  + No support for "flags"
  + Binding to EP is unconditional (ignoring flags)

Essentially this call currently just allows a client to split the
segment creation step from other initialization.  This is valuable in
that it allows a client to communicate via AMs (Short and Medium only)
prior to creating a segment.  While that is not tested directly by any
GASNet test at this point, there is a full gasnet_barrier() call at
the end of ClientInit() and most conduits use AMs for the exchange
within this call.

As with other recent work, the goal here is just to get something up
and running.  There is a hope that the boiler-plate nature of this
function will make for easy conduit-independent factorization at a
later time.  However, with so little of the intended final features
implemented, it would be premature to think we know how much
conduit-specific variation will be required.