tests: update testhello.c
This commit updates tests/testhello.c to initialize using the functions
gasnetex_ClientInit() and gasnetex_TeamSegmentCreate(), instead of
gasnet_init() and gasnet_attach() which are now only available to
clients which include gasnet.h.  Note that use of the EX interface for
handler registration has been in use for awhile already.