Update the gasnet testsuite with new APIs
This commit updates the GASNet testsuite to initialize using the
functions gasnetex_ClientInit() and gasnetex_TeamSegmentCreate(),
instead of gasnet_init() and gasnet_attach() which are now only
available to clients which include gasnet.h.  Note that use of the EX
interface for handler registration has been in use for awhile already.

This work removes the hidden "myteam" variable in test.h, making it an
explicit file-scope global in each test.

In the case of testinternal's supporting code in gasnet_diagnostic.c,
the team is now passed to gasneti_run_diagnostics() as an explicit
argument.

This commit also updates logic in test.h which creates a hidden test
segment when in EVERYTHING mode.  In particular it now intercepts
either gasnet_attach() or gasnetex_TeamSegmentCreate() depending on
whether the client test has included gasnet.h or gasnetex.h.  This was
previously the code that required the hidden "myteam" variable, but
that is now obtained from the arguments to the captured call to.
gasnetex_TeamSegmentCreate() (or as "g2ex_team" in the GASNet-1 case).