Fix most of bug 3528 - C++11 deprecates 'register'
This commit ensures that "register" will not be passed to a C++11 (or
later) compiler by the code in gasnet_atomic_bits.h.  There is one
special exception for ARM, which is explained in the comments.

A related change in AMUDP is pending.

(cherry picked from commit 728c4a8da11e8347e595fed555cc9359ee5643ce)