gasnet_map.c: move presegment back to hidden state
This commit moves the "presegment" used to carry state between calls
to gasneti_segmentInit() and gasneti_segmentAttach() back into
gasnet_mmap.c (as a file-scope static variable).  This, together with
a "legacy_mode" argument added to gasneti_segmentAttach(), results in
one less thing that each conduit was required to do exactly the same.