aries/gemini: implement IMMEDIATE AMRequest
This commit implements the IMMEDIATE flag in the following APIs:

  + gasnetex_AMRequestShortM
  + gasnetex_AMRequestMediumM
  + gasnetex_AMRequestLongM

The implementation is such that the Short and Medium AMRequest
functions "fail" (return 1) if and only if one of the following
conditions is true:

  + No post descriptor available for AM header
  + Another thread holds the per-peer buffer allocator lock (PAR only)
  + Per-peer remote buffer allocation fails (equiv to credits)
  + Reached limit on reply tracking data (GASNET_NETWORKDEPTH_TOTAL)

There are two distinct cases for Longs: packed and non-packed.

The packed case occurs when the payload is small enough to send with
the header like a Medium.  For this case, the IMMEDIATE support is
exactly as given above for Short and Medium.

The non-packed case uses an RDMA Put for the Long payload.  In this
case IMMEDIATE failure occurs if and only if there is no post
descriptor available for the payload Put.  Once the Put is launched
the Long is obligated to run even if it would stall for any one of the
reasons an IMMEDIATE Short or Medium might "fail".