gasnetrun_*: improve a failure mode
Provide more understandable error behavior if the user specifies a
valid file to gasnetrun_* but appears to lack execute permissions.

This arose on JLSE where the filesystem permissions for the KNL nodes
behave "strangely" and caused me significant confusion until I
realized it was a permissions problem.

It's worth noting perl's -x operator is apparently more conservative
(and less reliable) than the shell's test -x. perl 5.16.3 was seen to
trust the permissions it interprets from the directory listing and
return an (incorrect) negative answer on a batch node with incomplete
UID information, whereas actually attempting to open the file for
executing works (and shell test -x returns the correct answer).