Assert that client catches IMMEDIATE failures
The expectation we have for the IMMEDIATE flag is that the client
must always examine the return value.  In the case of explicit-handle
operations this includes checking for GASNETEX_NO_OP_HANDLE.  So, this
commit adds an assertion intended to catch cases in which the client
has allowed this value to reach a test or wait call.