Remove the cached segment upper-bound arrays
At Dan's suggestion, this commit removes the "_ub" arrays which had
been used to cache the ending address of the aux and client segments.
Not only does this eliminate one O(procs) data structure, but it also
may *reduce* the cost of the bounds-checking it was intended to
support (the upper-bound array replaced a single word-sized ADD of a
value already in cache with a load of a location that might *not* be
in cache - probably not the right trade-off for modern processors).