ibv: repair regressions in --disable-mmap support
This commit restores support in ibv-conduit for --disable-mmap, which
was broken by the recent work on segment creation.

+ Must unpin, but not unmap, auxseg at shutdown (same as client seg)
+ Cannot use gasneti_mmapLimit() without mmap() support

While use of InfiniBand w/o mmap() is not a likely production target,
a concious choice was made to support this configuration in order to
provide means to exercise the non-mmap code with a native conduit.