ibv: repair regressions in --disable-mmap support
This commit restores support in ibv-conduit for --disable-mmap.
Specifically, one cannot use gasneti_mmapLimit() without mmap()
support.  Additionally, this commit factors the code that does make
that call, as well as the call to gasneti_segmentInit(), eliminating
pointless duplication.

While use of InfiniBand w/o mmap() is not a likely production target,
a concious choice was made to support this configuration in order to
provide means to exercise the non-mmap code with a native conduit.