aries/gemini: implement IMMEDIATE AMReply
This commit implements the IMMEDIATE flag in the following APIs:

  + gasnetex_AMReplyShortM
  + gasnetex_AMReplyMediumM
  + gasnetex_AMReplyLongM

The implementation is such that the Short and Medium AMReply
functions "fail" (return 1) if and only if one of the following
conditions is true:

  + No post descriptor available for AM header
  + A bounce-buffer is needed but none is available

As with AMRequest, the AMReplyLong implementation of IMMEDIATE depends
on the size of the payload.  For payloads small enough to send with
the header like a Medium, the IMMEDIATE failure cases are those given
above.  For later payloads IMMEDIATE failure occurs if and only if
there is no post descriptor available for the payload Put.  Once the
Put is launched the Long is obligated to run even if it would stall
for any one of the reasons an IMMEDIATE Short or Medium might "fail".