pami: use PAMI_Send_immediate() for small Med/Long
This commit makes a long-overdue enhancement to the AM Medium and Long
paths to use PAMI_Send_immediate() when the sum of the header and
payload sizes is small enough permit it (128 bytes on BG/Q).

Shorts have always used PAMI_Send_immediate unconditionally.

This shows a measurable reduction in AM Req/Rep ping-pong latency for
payloads which fit.  Roughly a 5% reduction (from 6.3 to 6.0us) for
16 byte Mediums, and 25% (from 8.9 to 6.7) for 16 byte Longs.

P.S. This work was begun as a first step toward exposing local
completion support in EX, but it was clear that this work was
beneficial in GASNet-1 as well.