aries/gemini: ensure compliant LC implementation
This commit replaces the two unimplemented local completion options in
PutNB and PutNBI, previously both calling gasneti_fatalerror(), with
a "goto" the non-bulk (aka EVENT_NOW) implementation.

The reasoning behind the choice of "goto" is two-fold:

 + Ensures the proper if-ladder is retained
 + Is "cringeworthy" enough to help motivate proper implementations

It is worth noting that the original code in gasnete_put_nb() was
(incorrectly) cloned from the NBI case and so was diagnosing PTR as
invalid and GROUP as unimplemented, when their true roles are the
reverse.  So, the change to test gasneti_leaf_is_pointer() first in
gasnete_put_nb() is intentional.