pami: ensure compliant LC implementation
This commit makes two changes to ensure pami-conduit provides a
compliant implementations of the new local-completion options.

First, the two unimplemented local completion options in PutNB and
PutNBI, previously both calling gasneti_fatalerror(), now "goto" the
non-bulk (aka EVENT_NOW) implementation.  As with the same fix made in
aries/gemini, the choice of "goto" is intended to be "cringeworthy"
enough to help motivate proper implementations, and to retain the
proper conditional structure.

The gasnete_put_nbi() code was testing the wrong condition and so
incorrectly reported PTR as unimplemented and GROUP as invalid, when
their true roles are the reverse.  So, the change to test EVENT_GROUP
first in gasnete_put_nbi() is intentional.

Second, the AM Request and Reply functions for Medium and Long were
previously doing nothing with respect to local completion, which is
not correct for the case of a pointer-to-event as the lc_opt argument.
So, this commit adds the required gasneti_leaf_finish() to ensure the
client will see the invalid event required to indicate synchronous
local completion.