Empty all source files for non-EX conduits
This commit "removes" the contents of all source files (.c and .h) in
the five conduits that are not yet being ported forward to GASNet-EX
as part of our initial work.

All files remain in both the repository and the distribution tarball
to avoid the need to remove (and restore) the corresponding autoconf
and automake logic.  However, the source files now each contain just
an explanatory comment.  This is intended to reduce confusion for
anyone examining the sources, as well as ensuring that any git commits
cherry-picked from develop *will* have conflicts in those directories.

Prior to this commit, these five conduits contained only *partial*
updates with respect to GASNet-EX infrastructure changes and had also
not been receiving updates and bug fixes made on develop.