Move internal idiotcheck to gex_Client_Init()
This commit updates the mechanism by which we detect/report clients
which have erroneously included gasnet_internal.h. The previous
approach was to capture calls to gasnet_attach(), while the new logic
instead captures gex_Client_Init().

Additionally, we now catch the inclusion of gasnetex.h and
gasnet_internal.h in the same compilation unit at *link* time, as
opposed to printing an error message at run time.

This change preserves the name-shifting of gex_Client_Init() encoding
the version and configuration information in order to catch linking
errors (mixing of .o / .a from different versions or configurations).

There is no longer any need to capture gasnet_attach() because it is
defined as an inline in gasnet.h and we already preclude inclusion of
gasnet.h and gasnet_internal.h in the same compilation unit (in either
order, though the precise failure modes differ).