Begin outlining the four basic types
This commit provides definitions of the conduit-independent portions
of the four basic object types in GASNet-EX:

   + gex_Client_t
   + gex_Segment_t
   + gex_EP_t
   + gex_TM_t

There is a level of separation between the client-visible types,
above, and the internal types such as gasneti_client_t.  This is
implemented with the aid of gasneti_{import,export}_[type]() inline
functions which provide mechanisms for data integrity (magic numbers)
and enforcement of the abstraction barrier (swizzling of the client's
pointers) in a DEBUG build while reducing to just the appropriate type
cast in an NDEBUG build.

The implementation of each type is subject to conduit-specific
override, but the default provides the conduit-independent fields and
accessors for them.  These accessors directly access the internal data
structures, avoiding the costs of a function call.