Ensure gasnet_init() creates base objects
This commit moves some (stub) logic from gasnetc_attach_primary() to
gasnetc_Client_Init().  This change will ensure that once the stub are
fleshed-out, a call to the legacy gasnet_init() will result in the
(hidden/thunk) client, endpoint and team fully initialized.  That, in
turn, is necessary if gasnet_s{mynode,node}() are to be implemented in
terms of gex_TM_Query{Rank,Size}() and remain legal between
gasnet_init() and gasnet_attach().