Allocate real objects in Client_Init()
This commit fleshes out stub code in gasnetc_Client_Init() to perform
allocation and initialization of the client, endpoint and team_member
objects.  Additionally, testgasnet is extended to test all of the
currently defined accessors on the three corresponding objects.

NOTE: this code updated in this commit is a candidate for
factorization, but we aren't quite there yet.