ibv: fix broken blcr+gasnet-verbose
This commit removes calls to gasnet_get() and gasnet_put() which were
used with --enable-gasnet-verbose to test RDMA functionality right
after a checkpoint.  The calls are commented out with a note about
either removing them or replacing them.  In the meantime, there is
space after gasnet_ to avoid interference with identifier audits.