Updates to NBI and Event test/wait interfaces
This commit makes the following changes based on resolutions in the
conference calls held on 2017.06.12 and 2017.06.13:

+ Rename: GEX_NBI_{GETS,PUTS,ALL} -> GEX_EC_{PUT,GET,ALL}
+ Rename: GEX_NBI_LC -> GEX_EC_LC
+ Add: GEX_EC_AM
+ Remove: gex_NBI_{Test,Wait}{Puts,Gets,All}()
+ Add flags argument to: gex_Event_{Test,Wait}{All,Some}()

In addition to those changes, there is a correction to the
implementation of gex_NBI_{Test,Wait}(GEX_EC_PUT,...) which previously
would have erroneously allowed "sync" of the ROOT Put without its
subordinate LEAF LC (though the now-removed gex_NBI_{Test,Wait}Puts()
had this correct).  This involved a change to the values of the
(renamed) GEX_EC_* constants to encode to which IOP counters were to
be synced.

A large fraction of this commit us dedicated to updating the tests to
conform to these interface changes.