Fix bug 3560 - implement IMMEDIATE in AMPSHM
This commit adds support for GEX_FLAG_IMMEDIATE in AM Request and
Reply injection over PSHM.  In both Request and Reply cases, the call
may "fail" if another thread holds the lock on the allocator, or if
the allocator returns failure on its first attempt.

This commit also removes use of GASNETI_RETURN() from the AM Request and
Reply function in smp-conduit, since with the addition of IMMEDIATE
support to AMPSHM there is now a non-error case of non-zero return.