(Re)fix bug 3560
This commit repairs AM injection code in three conduits to account for
the recent addition of IMMEDIATE support in AMPSHM with the
corresponding possibility of a non-zero return value which is NOT an
error.

In the cases of udp- and mpi-conduits we continue to check for the
possibility that the underlying AMX routine has returned an error
code.  In the case of pami-conduit all errors at the PAMI level are
currently fatal.  There are no more branches than before, and in
some cases the critical paths are now actually one branch shorter.