Add AMMax*() symmetry test
Send AM's to every peer to check that AMMax*() return symmetric
results for the same set of arguments in both directions along
every path. This could potentially be replaced with an exchange
collective for better scalability, but the naive implementation
has the benefit of ensuring the AM network is fully connected.

This commit also exercises a second call to EPRegisterHandlers,
and utilizes don't care handler registration.