Implement gasneti_AMPSMH_TokenInfo()
This commit implements the "baseline" TokenInfo query in AMPSHM and
updates the still-evolving boilerplate in all conduits accordingly.
In doing so, the token is now instantiated on the stack in all build
(previously DEBUG had it on the heap and NDEBUG had no memory for the
token).

Fields have been added to the token to support queries on Req/Rep and
Short/Medium/Long once those are defined.

Also fixes some minor issues with the pshm/conduit token interface.
This has the effect of breaking the (never used) capability for
conduits to override the AMPSHM token.  That could be restored with
some effort if needed, but doing so is difficult without a full
override implemented.