Move AM handler tables into the EP object
This commit moves the storage of the AM handler tables into each EP
object.  This is an initial step towards multi-EP support.

While the storage and initialization of the tables has changed, the
AM receive paths in all conduits remains unchanged - still using the
global gasnetc_handler, which has changed type from array to pointer.
Over time, conduits are anticipated to identify the proper EP from the
incoming message as the means to locating the correct AM handler
table, yielding multi-EP support and eliminating this global variable.

In the case of ibv and gemini/aries conduits there are native
bootstrap collectives which are constructed over AMs.  By their nature
these are used before the point where the template code for
Client_Init() creates the first Endpoint.  For this reason, these
conduits now create the first Client and Endpoint within
gasnetc_init(), rather than later in its caller, Client_Init().  In
time, it is possible that reworking of one or both conduits for full
multi-EP support may change the way this is structured.  However, for
now this provides a clean solution to AM-based bootstrap collectives
in these conduits while minimizing variation from the (still evolving)
templates for Client_Init() and EP_Create().

The template for Client_Init() has also changed, in part to remove
some out-of-date TODO annotations, but more significantly to make
clear the differences between the first call and any subsequent ones.