fixup^2: Implement new negotiated-payload AM APIs
This commit corrects an "thinko" in the previous commit.

The logic to replace lc_opt==NULL with GEX_EVENT_NOW for GASNet-owned
buffers was conditioned on sd->_tofree.  However, another portion of
the same commit had already identified _tofree as being NULL for the
case of a GASNet-owned *zero-byte* buffer.  The work-around for that
issue was to allocate 1 byte in a DEBUG build.  However, NDEBUG builds
would continue to pass lc_opt==NULL for zero-byte GASNet-owned
buffers.

Commit now replaces *any* instance of lc_opt==NULL with GEX_EVENT_NOW,
knowing that an lc_opt of NULL is only valid for a GASNet-owned buffer
(and a DEBUG build would fatalerror 1-line earlier otherwise).