Opaque types need not be pointer-to-struct
This commit acts on a resolution on the 2017.06.26 conf call to
remove the forward struct declarations and replace by "..." the
instances of those structs in typedefs for the following types:

  + gex_Event_t
  + gex_Token_t
  + gex_EP_t
  + gex_Client_t
  + gex_Segment_t
  + gex_TM_t
  + gex_AM_SrcDesc_t

This change is reflected in both the .spec and in testgaset.h (which
no longer asserts that these type are pointers, nor initializes any
variables of these types to NULL).  Correspondingly, the macro
CHECK_NULL_CONSTANT() in testgasnet is renamed CHECK_ZERO_CONSTANT().