ibv: reorder AM handler registration
This commit makes a series of changes related to the order of AM
handler registration.  These are small steps toward being able to move
the AM handler table into the EP, which in turn will be needed for
(the most likely implementation of) multi-EP support.

1. The "early" registration of core handlers is reduced to just the
two required for bootstrap collectives, and is moved slightly later in
gasnetc_init() to avoid splitting the handling of LIDs (HCA
addresses).

2. The "normal" (not early) registration of core handlers, which was
previously disabled by "#if 0", is re-enabled and a typo fixed.

3. The registration of firehose AM handlers is relocated to
EP_Create() along with the other handlers, rather then being early.

There are no "meaningful" logic changes in the work described above,
just changes in the ordering of existing registration operations.
However, there is one minor logic change required as a result which is
to move the call to gasneti_amtbl_check() one statement later to avoid
applying it to the index==0 "ack" case (which may now occur before
that slot is initialized).