Implement gex_System_QueryJob{Rank,Size}()
This commit implements and documents

 + gex_System_QueryJobRank() to replace gasnet_mynode()
 + gex_System_QueryJobSize() to replace gasnet_nodes()

The GASNet-1 interfaces are relegated to "thunk" status (only defined
in gasnet.h), which necessitates updates to the tests which include
gasnetex.h instead of gasnet.h.

A pair of checks in testgasnet.h to compare the GASNet-1 APIs against
queries on team0 have been updated to the new APIs, and a pair of
checks added to testlegacy.c comparing the old and new APIs
head-to-head.

In docs/GASNet-EX.spec the changes include moving the (brief) existing
documentation of gex_Rank_t ahead of the description of Events, and
replacing the description of gex_Token_Info_t's source rank with a
reference to gex_System_QueryJobRank() (which contains the old text).

This commit does not introduce the agreed-upon "JobLive" query since
it feels too early to start adding resilience-centric APIs (even if
this one is trivial).