ibv: rewrite handling of loopback and PSHM AMs
This commit rewrites ibv-conduit's handling of AMs over both
conventional loopback and PSHM.  This yields measurably improved
latencies (such as loopback ping-pong times reduced from 82 to 66ns).

The old code tried to force loopback AMs to appear as if they had
arrived from the network.  This resulted in less source code to
maintain, but added complexity in both the AM injection and AM receive
paths.  It also resulted in substantially different handling than the
PSHM case (added later).

The handling of PSHM and true loopback are now in a common framework
that is distinct from the path for AMs arriving from the network.  The
result is that all AM dispatch paths are shorter, with the loopback
one being the most noticeably improved.