gemini/aries: fix a typo
This commit fixes a typo (present since Sep 2013) which caused
loopback (non-PSHM) Long Reply handlers to receive a token which
incorrectly identified the context as a Request.  This was never
visible to the client, and has not been seen to lead to errors.

(cherry picked from commit d7585631d891f5dd24a8fb1cde687bcb316b1371)