aries/gemini: refactor loopback and PSHM AMs
This commit follows the pattern of ibv-conduit's recent updates to AM
handling to place PSHM-based loopback and "pure loopback" into the
same framework.  Unlike the ibv-conduit case these changes are
intended to improve maintainablity of the code rather than improve
performance (which remains unchanged since the generated code should
be identical - or nearly so).