pami: factor PSHM/loopback AM logic
This commit follows the pattern of ibv and aries/gemini conduits'
recent updates to AM handling to place PSHM-based loopback and "pure
loopback" into the same framework.  Like the gemini/aries changes
these change are intended to improve maintainablity of the code rather
than improve performance (since the generated code should be nearly
identical).  For instance, this halves (from 6 to 3) the number of
instances of the loopback support that will need to change to add a
"real token" to pami-conduit.