Update/improve gex_Token_Info() implementations
This commit updates all implementations of (and template for)
gex_Token_Info() to set all supported fields of the output and *not*
branch on the bits in the input 'mask'.

Additionally, as suggested in the comments of bug 3574, this commit
adds DEBUG-only conduit-independent logic to post-process the output
to invalidate any fields *not* requested in the client's 'mask'.  The
intent is to make it less likely that clients can successfully utilize
fields they did not request.  Otherwise a client might unintentionally
depend on the new NDEBUG behavior of initializing all supported fields
(a behavior conduits are *not* required to follow).

The post-processing logic actually goes a little further than was
suggested in bug 3574 to perform some additional checks:

 + Input 'mask' must not contain any unrecognized bits
 + Conduit must produce values for any *required* fields in 'mask'
 + Requested fields must contain plausible results