ibv: start propagating EP through AM paths
This commit begins the process of threading the EP info through the AM
injection and reception code paths of ibv-conduit.  This work is
sufficient to eliminate the global table of AM handlers and perform
all AM dispatch in terms of the EP.

Unfortunately, removal of the global AM table is offset by the
addition of a global "ep0" which is required in some paths that do not
yet (or perhaps ever) have access to the EP from their callers.  It is
also used in numerous assertions to ensure EP values are not
uninitialized.

In this commit the conduit-specific EP data structure is created but
does not yet differ from the conduit-independent EP.  However, it will
in the future and creating it now avoids the need to change the types
of numerous function arguments later.