ibv: work around (potential) bug 1586
This commit replaces two instances of "static volatile int" with just
"static int" and volatile casts on each reference.  This is required
to avoid bug 1586, a known optimizer bug in the PathScale compiler.

It is uncertain if this is strictly required, but the code generated
by most compilers is going to be nearly identical.