New object definitions per 2017.07.18 call
This commit restructures the definitions of the heavyweight object
types to ease construction of conduit-specific data types which
contain them as a prefix (w/o the need to literally prefix the
accesses such as "p->common._flags").  This includes the addition of
an "alloc_size" argument to the gasneti_alloc_[TYPE]() functions to
permit a conduit to request allocation of more space than the
conduit-independent object requires.

Additionally, the definitions of the MAGIC constants are moved to
gasnet_internal.h to reduce the risk of a client doing something
inappropriate with them.