Adjust spawn dependency on HAVE_GETIFADDRS
HAVE_GETIFADDRS is required by AMUDP spawned workers to support
the $*_WORKERIP user environment setting.  Previously, the dependency
was enforced on the master (who issued a warning when it was missing).
However this could result in confusing behavior when performing a
cross-platform launch using amudprun-host, where the master and slave
OS's differ in their level of IFADDR support.

IFADDR support is now instead verified by the slaves, who issue a
warning when appropriate.

This also has the nice side-effect of removing amudprun-host's only
behavioral dependence on IFADDR support.