ofi: Fix possible issues with initialization logic
The conduit up to this point has been modifying the fi_info structure
directly to indicate that certain endpoints will be RMA or MSG only. The
providers we have tested on so far haven't cared about this. However, the actual
way that this should be done is by modifying the hints structure and requerying
fi_getinfo() to let the provider fill the info fields as it needs. This commit
does just that.

This commit also moves up the address vector binding to before the endpoints are
enabled. This is a requirement from the manpages (but psm2/sockets don't
complain about it). The address lookup and insertion remains where it was,
after enabling the endpoints.