fix GASNET_ENV_DEFAULT_HELPER for BSD expr
BSD expr was generating warnings when configure was passed arguments like:
  --with-multiconf=+whatever   or  --with-multiconf=-whatever
because +/- are treated specially when they appears at the start of an expr
argument and we wanted them to be treated as part of opaque strings for pattern
matching.

Adopt the same fix used by autoconf of prefixing with a X character to
ensure our string arguments to expr are treated as strings.

(cherry picked from commit dfbfdc2d8e108e1b9e52ad5276e718845c115133)