Implement "foreign free" of eop and iop
This commit ensures that the single-event and array-of-event paths for
reap of eops and iops (from NBI access regions) will correctly handle
the case that an op was created by a different thread.

In order to enable this, we now define GASNETE_INVALID_THREADIDX as
the all-ones pattern of type gasnete_threadidx_t.  We anticipate that
this will be used more in the future.

This commit also adds testing of the new capability to testinternal
(via gasnet_diagnostic.c).