bug 3575: clarification to GASNet spec 1.8
As discussed in bug 3575, this spec correction clarifies that
"loopback" memory-to-memory contiguous put/get operations prohibit
any kind of overlap between the source and dest regions, otherwise
the resulting value is undefined (indeterminate).

This matches the behavior of current and past implementations that
use POSIX memcpy() for this case (instead of e.g. memmove()). Many
POSIX memcpy() implementations happen to result in no change to
memory when the source and destination pointers are exactly equal,
but the behavior for that case is not guaranteed by POSIX, and
this clarifies that ambiguity also applies to GASNet.