Fix a latent bug in debug mallocator OOM path
GASNet's debug mallocator (--enable-debug or --enable-debug-malloc)
prints an explanatory message when an allocation attempt fails
(ie system malloc() returns NULL) usually as a fatal error console message.

The code path that constructed that error message had a bug (dating back
to at least 2008) that would result in garbled output for the source
location included in that message, or (in rare conditions)
possibly even a SEGV before message output. This bug only affected
the out-of-memory path (which in most cases leads to a fatal error).