pgi: work around bug 3587
This commit works around bug 3587 which is causing incorrect assertion
failures in some nightly builds with the PGI compilers.  The approach
is simply to omit the assertion when building with pgcc, under the
assumption that since it is not testing a compiler-specific property
any violation would be caught "sooner or later" with another compiler.