Deprecate psm-conduit
This commit makes changes to deprecate psm-conduit, including a
prominent NOTICE at the top of psm-conduit/README, items in the
ChangeLog, and a warning from configure if psm-conduit support is
enabled.

In addition to deprecation of psm-conduit, this commit also makes
updates to ofi-conduit/README to *recommend* its use on Omni-Path.

Finally, this commit adds text to ofi-conduit/README to describe the
problems that can occur in hybrid MPI+GASNet applications with (at
least) libfabric's psm2 provider.  This is not new, but the
recommendation of ofi-conduit for use with Omni-Path makes mention of
this issue more important.