Chapel CI: expand tests to include HPCC and ISX
This commit greatly expands the coverage of the Makefile for the
Chapel CI testing, but initially only adds the HPCC (excluding stream)
and ISX benchmarks.

The command line options match those of Chapel's test runs, with the
exception that we do not suppress printStats for HPCC.  This
difference allows the capture of BenchmarkResults for these tests.