Chapel CI: kludge support for additional conduits
The Chapel docs don't list OFI, PSM or PAMI in their table of
launchers, but do list them among GASNet conduits.  This leaves their
support status unclear.  The status is particularly unclear for
psm-conduit due to a runtime/src/launch/ganetrun_psm directory which
is not the default for CHPL_COMM_SUBSTRATE=psm.  This is the subject
of Chapel github issue 7061.

To permit testing of these three conduits, this commit sets
CHPL_LANUCHER as appropriate and actually *creates* the required
Chapel launcher directory as a minor delta from an existing one if
needed.